requirements.txt
pandas==2.0.3
requests==2.31.0
psycopg2-binary==2.9.9
dvc[s3]==3.42.0
python-dotenv==1.0.0
